[graph_trainer] Refactor passes.py into focused modules
Split the 1037-line passes.py into focused modules:
- memory_policy.py: SAC tagging and registered memory policies
- inductor_passes.py: regional/full Inductor compilation + FlexAttention annotation
- cudagraph.py: moved cudagraph_pass and insert_kernel_annotations_pass here

passes.py remains as the orchestration layer (~426 lines) with pass
list construction, apply_graph_passes, and registries.
