+++ torchtitan/experiments/graph_trainer/memory_policy.py
@@ -1,0 +1,284 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""
+Memory policy passes for graph_trainer.
+
+Selective activation checkpointing (SAC) tagging and memory policy dispatch.
+Each saved forward activation can independently be tagged as MUST_SAVE,
+MUST_RECOMPUTE, or MUST_CPU_OFFLOAD.  The ``tag_with_memory_policy_pass``
+entry point selects a tagging strategy via ``--compile.memory_policy``.
+"""
+
+from __future__ import annotations
+
+import operator
+from collections import defaultdict
+from collections.abc import Callable
+
+import torch
+from torch.utils.checkpoint import CheckpointPolicy
+
+from torchtitan.distributed.activation_checkpoint import _get_save_ops
+from torchtitan.distributed.fsdp import get_fsdp_reshard_after_forward_policy
+from torchtitan.experiments.graph_trainer.common_utils import (
+    _get_layer_id,
+    _is_backward_node,
+    _MODULE_FQN,
+    _NOT_IN_LAYERS,
+)
+from torchtitan.experiments.graph_trainer.cpu_offload import (
+    tag_all_offloadable_activations,
+)
+from torchtitan.experiments.graph_trainer.log_activation_memory_policy import (
+    log_activation_memory_policy,
+)
+from torchtitan.tools.logging import logger
+
+
+# ---------------------------------------------------------------------------
+# Registry
+# ---------------------------------------------------------------------------
+
+MEMORY_POLICY_REGISTRY: dict[str, Callable] = {}
+
+
+def register_memory_policy(key: str):
+    """Decorator that registers a memory policy function."""
+
+    def decorator(fn: Callable) -> Callable:
+        MEMORY_POLICY_REGISTRY[key] = fn
+        return fn
+
+    return decorator
+
+
+def _make_default_memory_policy(
+    save_ops: set | None = None,
+    *,
+    fsdp_reshard_after_forward: bool = True,
+) -> Callable:
+    """Create a SAC policy function from a set of op targets to save."""
+    if save_ops is None:
+        save_ops = _get_save_ops()
+        if not fsdp_reshard_after_forward:
+            save_ops.add(torch.ops._c10d_functional.all_gather_into_tensor.default)
+
+    def policy_fn(node: torch.fx.Node) -> CheckpointPolicy:
+        if node.target in save_ops:
+            return CheckpointPolicy.MUST_SAVE
+        return CheckpointPolicy.PREFER_RECOMPUTE
+
+    return policy_fn
+
+
+def _make_eager_memory_policy(save_ops: set | None = None) -> Callable:
+    """Eager-compatible SAC policy that alternates mm ops between save/recompute.
+
+    Matches the behavior of torchtitan.distributed.activation_checkpoint:
+    every second mm/linear op is marked PREFER_RECOMPUTE instead of MUST_SAVE.
+    """
+    if save_ops is None:
+        save_ops = _get_save_ops()
+    mm_ops = {torch.ops.aten.mm.default, torch.ops.aten.linear.default}
+    mm_count = 0
+
+    def policy_fn(node: torch.fx.Node) -> CheckpointPolicy:
+        nonlocal mm_count
+        if node.target in mm_ops:
+            mm_count += 1
+            if node.target in save_ops and mm_count % 2 == 0:
+                return CheckpointPolicy.PREFER_RECOMPUTE
+        if node.target in save_ops:
+            return CheckpointPolicy.MUST_SAVE
+        return CheckpointPolicy.PREFER_RECOMPUTE
+
+    return policy_fn
+
+
+def apply_sac_pass(
+    gm: torch.fx.GraphModule,
+    example_inputs: tuple | None = None,
+    *,
+    policy_fn: Callable[[torch.fx.Node], CheckpointPolicy] | None = None,
+) -> torch.fx.GraphModule:
+    """Apply selective activation checkpointing on the joint graph.
+
+    Annotates forward ``call_function`` nodes with a ``CheckpointPolicy``
+    determined by ``policy_fn``. After tagging, a boundary pass forces
+    ``MUST_SAVE`` on recomputable nodes whose output crosses a layer
+    boundary (layer N → layer N+1), since recomputing them would require
+    rerunning the entire preceding layer.
+
+    ``getitem`` / ``wait_tensor`` nodes inherit the parent's tag.
+
+    The model must have been annotated with ``annotate_module_fqns`` before
+    tracing so that nodes carry ``module_fqn`` metadata.
+
+    Args:
+        gm: The joint forward-backward graph module.
+        policy_fn: Callable that takes a node and returns a CheckpointPolicy.
+            Defaults to ``_make_default_memory_policy()`` if None.
+
+    Returns:
+        The annotated graph module
+    """
+    if policy_fn is None:
+        policy_fn = _make_default_memory_policy()
+
+    layer_stats: dict[int, dict[str, int]] = defaultdict(
+        lambda: {"save": 0, "recompute": 0}
+    )
+
+    # Pass 1: Tag each forward node with a recompute policy.
+    for node in gm.graph.nodes:
+        if node.op != "call_function":
+            continue
+
+        # Skip backward nodes — they must not carry recompute tags,
+        # otherwise the remat pass would try to duplicate backward ops.
+        if _is_backward_node(node):
+            continue
+
+        # Skip the post-layer epilogue (lm_head + loss). Chunked-loss
+        # regions split backward into multiple disjoint regions, and the
+        # remat pass only supports one region with must_recompute deps.
+        fqn = node.meta.get("custom", {}).get(_MODULE_FQN, "")
+        if fqn.startswith(("lm_head", "loss")):
+            continue
+
+        if node.target in (
+            operator.getitem,
+            torch.ops._c10d_functional.wait_tensor.default,
+        ):
+            # Propagate from parent: getitem extracts tuple elements,
+            # wait_tensor is tied to its async collective — both must
+            # share the parent's save/recompute decision.
+            parent = node.args[0]
+            if isinstance(parent, torch.fx.Node) and "recompute" in parent.meta:
+                node.meta["recompute"] = parent.meta["recompute"]
+            continue
+
+        layer_id = _get_layer_id(node)
+
+        # NOTE: The eager SAC policy (activation_checkpoint.py) alternates
+        # mm ops between MUST_SAVE and PREFER_RECOMPUTE. We omit that here
+        # because the alternating heuristic is arbitrary.
+        node.meta["recompute"] = policy_fn(node)
+        key = (
+            "save"
+            if node.meta["recompute"] == CheckpointPolicy.MUST_SAVE
+            else "recompute"
+        )
+        layer_stats[layer_id][key] += 1
+
+    # Pass 2: Force MUST_SAVE at layer boundaries. If a recomputable node
+    # feeds into a node in a higher layer, saving it is cheaper than
+    # recomputing the entire preceding layer.
+    def _is_recomputable(n: torch.fx.Node) -> bool:
+        return n.meta.get("recompute") in (
+            CheckpointPolicy.PREFER_RECOMPUTE,
+            CheckpointPolicy.MUST_RECOMPUTE,
+        )
+
+    boundary_saves = 0
+    for node in gm.graph.nodes:
+        if _is_backward_node(node) or not _is_recomputable(node):
+            continue
+        node_layer_id = _get_layer_id(node)
+        for user in node.users:
+            if (
+                not _is_backward_node(user)
+                and _is_recomputable(user)
+                and _get_layer_id(user) > node_layer_id
+            ):
+                node.meta["recompute"] = CheckpointPolicy.MUST_SAVE
+                boundary_saves += 1
+                break
+
+    gm.recompile()
+    logger.info("Applied selective activation checkpointing (SAC) graph pass.")
+    if boundary_saves:
+        logger.info(f"  Forced {boundary_saves} nodes to MUST_SAVE at layer boundaries")
+    for layer_id in sorted(layer_stats):
+        stats = layer_stats[layer_id]
+        label = "non-layer" if layer_id == _NOT_IN_LAYERS else str(layer_id)
+        logger.info(
+            f"  Layer {label}: "
+            f"{stats['save']} MUST_SAVE, "
+            f"{stats['recompute']} PREFER_RECOMPUTE"
+        )
+    return gm
+
+
+@register_memory_policy("default")
+def _default_memory_policy_pass(
+    gm: torch.fx.GraphModule,
+    *,
+    config: "GraphTrainer.Config",
+) -> torch.fx.GraphModule:
+    """SAC policy that saves all compute-intensive ops and FSDP all_gathers."""
+    fsdp_reshard_after_forward = get_fsdp_reshard_after_forward_policy(
+        config.parallelism.fsdp_reshard_after_forward,
+        pp_enabled=config.parallelism.pipeline_parallel_degree > 1,
+    )
+    policy_fn = _make_default_memory_policy(
+        fsdp_reshard_after_forward=fsdp_reshard_after_forward,
+    )
+    apply_sac_pass(gm, policy_fn=policy_fn)
+    return gm
+
+
+@register_memory_policy("eager")
+def _eager_memory_policy_pass(
+    gm: torch.fx.GraphModule,
+    *,
+    config: "GraphTrainer.Config",
+) -> torch.fx.GraphModule:
+    """SAC policy that alternates mm ops between save/recompute."""
+    apply_sac_pass(gm, policy_fn=_make_eager_memory_policy())
+    return gm
+
+
+@register_memory_policy("budget_limited_offload")
+def _budget_limited_offload_memory_policy_pass(
+    gm: torch.fx.GraphModule,
+    *,
+    config: "GraphTrainer.Config",
+) -> torch.fx.GraphModule:
+    """SAC + CPU offload: apply default SAC, then offload within budget."""
+    _default_memory_policy_pass(gm, config=config)
+    tag_all_offloadable_activations(
+        gm, cpu_budget_gb=config.compile.cpu_offload_budget_gb
+    )
+    return gm
+
+
+def tag_with_memory_policy_pass(
+    gm: torch.fx.GraphModule,
+    example_inputs: tuple | None = None,
+    *,
+    config: "GraphTrainer.Config",
+) -> torch.fx.GraphModule:
+    """Tag forward nodes with MUST_SAVE, PREFER_RECOMPUTE, or MUST_CPU_OFFLOAD.
+
+    The ``config.compile.memory_policy`` selects the tagging strategy:
+        default: SAC with all compute-intensive ops saved.
+        eager: SAC alternating mm ops between save/recompute.
+        budget_limited_offload: SAC + CPU offload within budget.
+
+    Other memory policies combining SAC and CPU offload can be added
+    via ``register_memory_policy`` without modifying this function.
+    """
+    memory_policy = config.compile.memory_policy
+    if memory_policy not in MEMORY_POLICY_REGISTRY:
+        raise ValueError(
+            f"Unknown memory_policy: {memory_policy!r}. "
+            f"Available: {list(MEMORY_POLICY_REGISTRY.keys())}"
+        )
+    gm = MEMORY_POLICY_REGISTRY[memory_policy](gm, config=config)
+    log_activation_memory_policy(gm)
+    return gm
